requirements.txt
Flask==3.0.3
requests==2.32.3
Unidecode==1.3.8
python-dotenv==1.0.1
pandas==2.2.3
matplotlib==3.9.2
selenium=4.26.1
webdriver-manager=4.0.2